Issue 27: Refactor DocumentsBundle contexts
